implementing delete page for all model in sidebar. using default delete component exept for documents

+++ app/controllers/servers_controller.rb
@@ -74,6 +74,11 @@ class ServersController < ApplicationController # rubocop:disable Metrics/ClassL
   end
 
   def destroy
+    unless params["confirm"] == "true"
+      render
+      return
+    end
+
     respond_to do |format|
       if @server.destroy
         format.html { redirect_to servers_path(search_params), notice: t(".flashes.destroyed") }
